Extract traverse function from V3 serializer to utils
